offload route data handling to new hook
